revert: downgrade sysinfo to v0.32
This reverts the sysinfo library from v0.37 to v0.32, because the
upgrade introduced a fault in function
sysinfo::unix::apple::cpu::get_cpu_frequency on arm64 architectures.
This error occurs within the process group termination feature that
sysinfo was introduced for.

Official changelogs of sysinfo for both v0.33 and 0.35 indicate
modifications related to CPU frequency functions, which are likely the
cause of the error. So a temporary revert is considered the safest
solution.