feat: add two types of soniox finalize mode
The default mode is "default", which is the same behaviour as before.
The mode "ignore" ignores the `asr_finalize` command, you should ensure
`enable_endpoint_detection` is enabled, because you will need to rely on
the <end> token to determine asr_finalize_end.
When the mode "mute_pkg" is set, `asr_finalize` will not call the
`finalize` API of soniox. Instead, a mute package sized by the config
`mute_pkg_duration_ms` is sent to simulate the `finalize` command. The
`asr_finalize_end` is sent immediately after the mute package is sent,
so it it unreliable.